+++ android/src/main/java/com/brentvatne/common/api/SubtitleStyle.kt
@@ -1,12 +1,21 @@
 package com.brentvatne.common.api
 
-import com.brentvatne.common.toolbox.ReactBridgeUtils
+import androidx.core.graphics.toColorInt
+import com.brentvatne.common.toolbox.ReactBridgeUtils.safeGetFloat
+import com.brentvatne.common.toolbox.ReactBridgeUtils.safeGetInt
+import com.brentvatne.common.toolbox.ReactBridgeUtils.safeGetString
 import com.facebook.react.bridge.ReadableMap
 
 /**
  * Helper file to parse SubtitleStyle prop and build a dedicated class
  */
 class SubtitleStyle private constructor() {
+    var foregroundColor: Int? = null
+        private set
+    var backgroundColor: Int? = null
+        private set
+    var windowColor: Int? = null
+        private set
     var fontSize = -1
         private set
     var paddingLeft = 0
@@ -21,6 +30,9 @@ class SubtitleStyle private constructor() {
         private set
 
     companion object {
+        private const val PROP_FOREGROUND_COLOR = "foregroundColor"
+        private const val PROP_BACKGROUND_COLOR = "backgroundColor"
+        private const val PROP_WINDOW_COLOR = "windowColor"
         private const val PROP_FONT_SIZE_TRACK = "fontSize"
         private const val PROP_PADDING_BOTTOM = "paddingBottom"
         private const val PROP_PADDING_TOP = "paddingTop"
@@ -29,15 +41,23 @@ class SubtitleStyle private constructor() {
         private const val PROP_OPACITY = "opacity"
 
         @JvmStatic
-        fun parse(src: ReadableMap?): SubtitleStyle {
-            val subtitleStyle = SubtitleStyle()
-            subtitleStyle.fontSize = ReactBridgeUtils.safeGetInt(src, PROP_FONT_SIZE_TRACK, -1)
-            subtitleStyle.paddingBottom = ReactBridgeUtils.safeGetInt(src, PROP_PADDING_BOTTOM, 0)
-            subtitleStyle.paddingTop = ReactBridgeUtils.safeGetInt(src, PROP_PADDING_TOP, 0)
-            subtitleStyle.paddingLeft = ReactBridgeUtils.safeGetInt(src, PROP_PADDING_LEFT, 0)
-            subtitleStyle.paddingRight = ReactBridgeUtils.safeGetInt(src, PROP_PADDING_RIGHT, 0)
-            subtitleStyle.opacity = ReactBridgeUtils.safeGetFloat(src, PROP_OPACITY, 1f)
-            return subtitleStyle
-        }
+        fun parse(src: ReadableMap?): SubtitleStyle =
+            SubtitleStyle().apply {
+                foregroundColor = runCatching {
+                    safeGetString(src, PROP_FOREGROUND_COLOR)?.toColorInt()
+                }.getOrNull()
+                backgroundColor = runCatching {
+                    safeGetString(src, PROP_BACKGROUND_COLOR)?.toColorInt()
+                }.getOrNull()
+                windowColor = runCatching {
+                    safeGetString(src, PROP_WINDOW_COLOR)?.toColorInt()
+                }.getOrNull()
+                fontSize = safeGetInt(src, PROP_FONT_SIZE_TRACK, -1)
+                paddingBottom = safeGetInt(src, PROP_PADDING_BOTTOM, 0)
+                paddingTop = safeGetInt(src, PROP_PADDING_TOP, 0)
+                paddingLeft = safeGetInt(src, PROP_PADDING_LEFT, 0)
+                paddingRight = safeGetInt(src, PROP_PADDING_RIGHT, 0)
+                opacity = safeGetFloat(src, PROP_OPACITY, 1f)
+            }
     }
 }